+++ Indexer/src/test/TestIndex.java
@@ -12,20 +12,20 @@ public class TestIndex {
     
     public TestIndex() {
         index = new Index(null);
-        index.addTerm("hola", url1, 3);
-        index.addTerm("prueba", url1, 1);
-        index.addTerm("documento", url1, 1);
-        index.addTerm("test", url1, 8);
+        index.addTerm("hola", url1, (short)3);
+        index.addTerm("prueba", url1, (short)1);
+        index.addTerm("documento", url1, (short)1);
+        index.addTerm("test", url1, (short)8);
         
-        index.addTerm("hola", url2, 5);
-        index.addTerm("caracola", url2, 3);
-        index.addTerm("que", url2, 3);
-        index.addTerm("haces", url2, 3);
+        index.addTerm("hola", url2, (short)5);
+        index.addTerm("caracola", url2, (short)3);
+        index.addTerm("que", url2, (short)3);
+        index.addTerm("haces", url2, (short)3);
         
-        index.addTerm("hola", url3, 1);
-        index.addTerm("que", url3, 2);
-        index.addTerm("prueba", url3, 3);
-        index.addTerm("test", url3, 4);
+        index.addTerm("hola", url3, (short)1);
+        index.addTerm("que", url3, (short)2);
+        index.addTerm("prueba", url3, (short)3);
+        index.addTerm("test", url3, (short)4);
         
         String[] termVector = index.getTermVector();
         for (String s : termVector) {
@@ -35,7 +35,7 @@ public class TestIndex {
         
         String[] docVector = index.getDocumentTermVector();
         
-        long[][] freqMatrix = index.getTermFreqMatrix();
+        short[][] freqMatrix = index.getTermFreqMatrix();
         for (int i = 0; i < docVector.length; i++) {
             System.out.print(docVector[i] + ": ");
             for (int j = 0; j < termVector.length; j++) {
